refactor : 오브젝트 드래그, 모델, 조합 컴포넌트 분리
