+++ Assets/Project/Prefabs/TimerBar.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a36329751055894ea18d15ad63bf013
+guid: d3b8ac0ab81c24541b3f06cc8f98d618
 PrefabImporter:
   externalObjects: {}
   userData:
