--- Assets/Project/Prefabs/Wood.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 781d6a1edb9c42f47a76ed6e1a8d5d2f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
