+++ Assets/Project/Scripts/Card/CardCombine.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 499b6faa96fb48a40b2406dbfeb753a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
