+++ Assets/Project/Scripts/Card/CardModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32ff8d72382bfdc428e2a9649fd657c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
