+++ Assets/Project/Scripts/Util.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 2143979037f874d41a21ed7426f4557b
-PrefabImporter:
+guid: 952c7917dffe6604bbb3834e1b117be9
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
